Remove the cacheDiffForKnownOperation observable array optimization
It was causing issues in DotVVM as it bypasses the observable validator.
The cacheDiffForKnownOperation needs to see the array and method
arguments before the validator is called, which means that unprocessed
objects might start leaking into the 'arrayChanges' subscribers.

Since DotVVM does not use the observableArray instance method
anywhere, the optimization cannot help significantly.
Array change tracking still works, but will fall back on the slower diffing algorithm.